+++ packages/@ionic/cli/src/commands/cordova/build.ts
@@ -92,7 +92,7 @@ The Cordova CLI requires a separator for platform-specific arguments for Android
         const runner = await this.project.requireBuildRunner();
         const runnerOpts = runner.createOptionsFromCommandLine(inputs, generateOptionsForCordovaBuild(metadata, inputs, options));
         await runner.run(runnerOpts);
-      } catch (e) {
+      } catch (e: any) {
         if (e instanceof RunnerException) {
           throw new FatalException(e.message);
         }