+++ packages/@ionic/utils-network/src/index.ts
@@ -16,10 +16,11 @@ function getDefaultAddresses(): string[] {
 
     for (const device of Object.keys(networkInterfaces)) {
       const networkInterface = networkInterfaces[device];
-
-      addresses.push(...networkInterface.map(i => i.address));
+      if (networkInterface) {
+        addresses.push(...networkInterface.map(i => i.address));
+      }
     }
-  } catch (e) {
+  } catch (e: any) {
     // swallow
   }
 
@@ -32,10 +33,11 @@ export function getExternalIPv4Interfaces(): NetworkInterface[] {
 
   for (const device of Object.keys(networkInterfaces)) {
     const networkInterface = networkInterfaces[device];
-
-    for (const networkAddress of networkInterface) {
-      if (!networkAddress.internal && networkAddress.family === 'IPv4') {
-        devices.push({ device, ...networkAddress });
+    if (networkInterface) {
+      for (const networkAddress of networkInterface) {
+        if (!networkAddress.internal && networkAddress.family === 'IPv4') {
+          devices.push({ device, ...networkAddress });
+        }
       }
     }
   }
@@ -81,7 +83,7 @@ export async function isPortAvailable(port: number): Promise<boolean> {
       if (!available) {
         return false;
       }
-    } catch (e) {
+    } catch (e: any) {
       debug('error while checking %s:%d: %o', address, port, e);
     }
   }
@@ -167,7 +169,7 @@ export async function isHostConnectable(host: string, port: number, { timeout }:
 
         resolve(true);
         resolved = true;
-      } catch (e) {
+      } catch (e: any) {
         // try again
       }
     }