+++ packages/@ionic/v1-toolkit/src/lib/log.ts
@@ -1,3 +1,0 @@
-import * as chalk from 'chalk';
-
-export const timestamp = () => chalk.dim(`[${new Date().toTimeString().slice(0, 8)}]`);