+++ packages/@ionic/cli-framework-output/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.6](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework-output@2.2.5...@ionic/cli-framework-output@2.2.6) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/cli-framework-output
+
+
+
+
+
 ## [2.2.5](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework-output@2.2.4...@ionic/cli-framework-output@2.2.5) (2022-06-16)
 
 **Note:** Version bump only for package @ionic/cli-framework-output