+++ packages/@ionic/utils-process/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.11](https://github.com/ionic-team/ionic-cli/compare/@ionic/utils-process@2.1.10...@ionic/utils-process@2.1.11) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/utils-process
+
+
+
+
+
 ## [2.1.10](https://github.com/ionic-team/ionic-cli/compare/@ionic/utils-process@2.1.9...@ionic/utils-process@2.1.10) (2022-06-16)
 
 **Note:** Version bump only for package @ionic/utils-process