+++ packages/cli-scripts/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.65](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.64...cli-scripts@2.1.65) (2023-03-29)
+
+**Note:** Version bump only for package cli-scripts
+
+
+
+
+
+## [2.1.64](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.63...cli-scripts@2.1.64) (2023-03-17)
+
+**Note:** Version bump only for package cli-scripts
+
+
+
+
+
+## [2.1.63](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.62...cli-scripts@2.1.63) (2023-03-17)
+
+**Note:** Version bump only for package cli-scripts
+
+
+
+
+
+## [2.1.62](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.61...cli-scripts@2.1.62) (2023-01-13)
+
+**Note:** Version bump only for package cli-scripts
+
+
+
+
+
+## [2.1.61](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.60...cli-scripts@2.1.61) (2023-01-13)
+
+**Note:** Version bump only for package cli-scripts
+
+
+
+
+
 ## [2.1.60](https://github.com/ionic-team/ionic-cli/compare/cli-scripts@2.1.59...cli-scripts@2.1.60) (2022-12-20)
 
 **Note:** Version bump only for package cli-scripts